utils/requirements.txt
tqdm
faker
psycopg2-binary
